LayoutTests/fast/frames/seamless/seamless-window-open-override.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
if (window.testRunner)
    testRunner.waitUntilDone();

window.addEventListener('message', function(evt) {
    shouldBeEqualToString("frames[0].frames[0].document.body.textContent", "DONE\n");

    if (window.testRunner)
        testRunner.notifyDone();
}, false);
</script>
<iframe srcdoc="
    Waiting...
    <iframe seamless
            srcdoc='
                <script>
                    window.open(&quot;resources/done.html&quot;,&quot;_self&quot;);
                </script>
            '>
    </iframe>
"></iframe>
